day 9 part 2

+++ 2018/9/src/Player.cs
@@ -6,14 +6,14 @@ namespace src
     class Player
     {
         public int Id { get; }
-        public int Score { get; private set; }
+        public long Score { get; private set; }
         
         public Player(int id)
         {
             Id = id;
         }
 
-        public void AddToScore(int value)
+        public void AddToScore(long value)
         {
             Score += value;
         }
